requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
python-decouple>=3.8,<4.0
psycopg2-binary>=2.9,<3.0
bcrypt>=4.1,<5.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.0,<22.0
redis>=5.0,<6.0
celery>=5.3,<6.0
Pillow>=10.0,<11.0
python-dateutil>=2.8,<3.0
reportlab>=4.0,<5.0
